flytt OpsjonRoutes til AvtaleRoutes
