+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/routes/ApiRoutes.kt
@@ -45,7 +45,6 @@ fun Route.adminflateRoutes() {
     brregVirksomhetRoutes()
     notificationRoutes()
     janzzRoutes()
-    opsjonRoutes()
     tilsagnRoutes()
 }
